### 2017-07-09 #312 @whedon
[PRE REVIEW]: arrgh: a Go interface to the OpenCPU R server system
**Submitting author:** @kortschak (<a href="http://orcid.org/0000-0001-8295-2301">Robert Daniel Kortschak</a>)
**Repository:** <a href="https://github.com/kortschak/arrgh" target ="_blank">https://github.com/kortschak/arrgh</a>
**Version:** v1.0.0

kortschak if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](https://github.com/openjournals/joss/blob/master/docs/reviewers.csv) have already agreed to review for JOSS and may be suitable for this submission.
