### 2017-08-12 #312 @kortschak
I am confused, the paper clearly states that the package is intended to be used for interfacing with R, which is primarily a research software. The paper also explains why this would be useful, in that Go is a good language for performing scientific analyses for many reasons, but that R provides a much richer statistical analysis support base and so use the two together is a benefit:

>Go is a simple statically typed compiled language that provides many benefits for scientific computing. However, Go currently lacks the rich statistical analysis tools available in R. So an interface between R and Go would allow building analytical tools utilising the simplicity and high performance of Go and the statistical analysis tools available within R.
