### 2017-08-12 #312 @kortschak
To give an explicit use case (hypothetical), if a person wanted to perform a genomics analysis using Go and the [bíogo ncbi package](http://joss.theoj.org/papers/10.21105/joss.00234) (interestingly built around the same server/client model as the arrgh package) requiring some statistical analysis, it may be necessary to use R. The arrgh package would facilitate this analysis. The need for statistical analysis to support research seemed so obvious to me that I did not include any text to justify it; should I?
