### 2017-08-14 #312 @pjotrp
@kortschak it will be good to give an example of use in research context. @arfon I can be the editor.
