### 2017-08-15 #312 @kortschak
I am teaching at the moment and this will take some time to put together, but it does honestly strike me as very odd. These are two languages that are used for scientific research, with different strengths that can, if interfaced, strengthen each other.
