### 2017-08-15 #312 @arfon
> I am teaching at the moment and this will take some time to put together, but it does honestly strike me as very odd. These are two languages that are used for scientific research, with different strengths that can, if interfaced, strengthen each other.

Hi @kortschak. We've been discussing this submission as an editorial board and the primary concern we have is that this submission falls into the 'utility package' category (http://joss.theoj.org/about#submission_requirements). We definitely understand the potential value of connecting these two languages, we're simply looking for an illustration of this for the reader. Better yet, if you have used this in some of your own ongoing research then cite that too.
