### 2017-08-15 #312 @pjotrp
@kortschak Dan, take a look at the RPy2 paper. https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-11-S12-S11
