### 2017-08-16 #312 @kortschak
@pjotrp I'm not sure what I should take from that. The two approaches are very different, reflecting the differences in the underpinning philosophies of python and Go, and also reflects the differences in approaches to performance optimisation seen in the two languages; the package use example of edger in figure 3 would be an anti-pattern in Go because of the heavy inter-process communication purely for the purpose of retaining the data in the python space. It is also difficult for me to distinguish the packages there from what @arfon describes here as a 'utility package'. The RPy2 packages do things that in my work would be done purely in R, and serve only to wrap that in a different syntax.

I will write up an example of a graph structure analysis where the graph statistics are generated in Go using packages in the gonum graph suite and the statistical analysis and plotting is done using arrgh to interface to R. If this plan is satisfactory, then I can move ahead, otherwise I am at a loss with how to move this forward, and would continue to be confused given other paper that have been moved onto review/published here recently that IMO sit squarely in the same space.
