### 2017-08-16 #312 @arfon
> I will write up an example of a graph structure analysis where the graph statistics are generated in Go using packages in the gonum graph suite and the statistical analysis and plotting is done using arrgh to interface to R. If this plan is satisfactory, then I can move ahead, otherwise I am at a loss with how to move this forward, and would continue to be confused given other papers that have been moved onto review/published here recently that IMO sit squarely in the same space.

Thanks @kortschak. I realize there many be some apparent inconsistencies in editorial policy - this is because we've recently updated our submission guidelines (see https://github.com/openjournals/joss/pull/307).
