### 2017-08-16 #312 @pjotrp
Hey Dan. We are all volunteers. Please appreciate that. All good intentions. I am with you that we should push forward. The idea of a GO graph is very good.
