### 2017-09-05 #312 @pjotrp
@kortschak ping me when you are ready and I'll look for a reviewer. Since you have published with JOSS before I trust you to be ready at some point.
