### 2017-09-05 #312 @kortschak
Thanks for the patience @pjotrp, I will let you know when I have the worked up examples as a PR.
