### 2017-12-15 297cc69
paper: add nuccore example
