### 2017-12-15 f178dc8
paper: add nuccore example
