### 2017-12-15 #312 @kortschak
@pjotrp *ping*

I have a [PR adding a more fleshed out example usage](https://github.com/kortschak/arrgh/pull/8).
