### 2017-12-19 #312 @pjotrp
@brentp would you mind reviewing this submission? Or suggest someone who would be interested in using R from Go?
