### 2017-12-19 #517 @whedon
[REVIEW]: arrgh: a Go interface to the OpenCPU R server system
**Submitting author:** @kortschak (<a href="http://orcid.org/0000-0001-8295-2301">Robert Daniel Kortschak</a>)
**Repository:** <a href="https://github.com/kortschak/arrgh" target ="_blank">https://github.com/kortschak/arrgh</a>
**Version:** v1.0.0

- [x] **Authorship:** Has the submitting author (kortschak) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
