### 2017-12-19 #517 @pjotrp
@brentp we review here. @kortschak would you be so kind to validate we can tick the boxes above? Thanks!
