### 2017-12-20 #517 @brentp
@pjotrp now that I read the COI doc, it appears that, by the letter, I would be in conflict. It says:
```
COIs are your present or previous association with any authors of a submission: recent (past four years) collaborators in funded research or work that is published...
```
I am on the [biogo/hts](http://joss.theoj.org/papers/10.21105/joss.00168) paper with @kortschak
