### 2017-12-20 #517 @kortschak
Sorry, I did not read that part - I focused on the areas relating to the paper/submission.
