### 2017-12-20 #517 @kortschak
May I suggest Jeroen Ooms. He is the author of OpenCPU which is either a good thing or a bad thing; he is the expert on the base system, but whether that constitutes a conflict here, I don't know. I have not worked with him.
