### 2017-12-20 #517 @pjotrp
@brentp I don't think being a co-author on an earlier paper is a conflict. @arfon can you advise?
