### 2017-12-20 #517 @danielskatz
The CoI policy is defined here: https://github.com/openjournals/joss/blob/master/COI.md and since the paper in question was earlier this year, this is a conflict.

Based on the following text:
> If you have a conflict of interest with a submission, you should disclose the specific reason to the submissions' editor. This may lead to you not being able to review the submission, but some conflicts may be recorded and then waived, and if you think you are able to make an impartial assessment of the work, you should request that the conflict be waived.

I think there are 2 questions:

@brentp,  do you feel you can make an impartial assessment of the work?

If so, then

@pjotrp, you should decide if you agree, and if so, declare the conflict recorded (here) and waived.

If no to either, then we need a new reviewer.
