### 2017-12-21 #517 @brentp
> @brentp, do you feel you can make an impartial assessment of the work?

yes
