### 2017-12-23 #517 @pjotrp
Thanks @danielskatz for stepping in. I feel confident in waiving this COI. @brentp please continue review.
