### 2018-01-08 kortschak/arrgh#9 @brentp
installation instructions
In order to install, I had to do
in ubuntu:
```
sudo apt-get install libprotobuf-c-dev
```

 in R:

```
install.packages('sys', configure.vars = 'NO_APPARMOR=1')
install.packages('opencpu')
```

I suspect this is not an issue for folks accustomed to using opencpu, but something like this might be informative in addition to linking to opencpu.

re: https://github.com/openjournals/joss-reviews/issues/517
