### 2018-01-08 kortschak/arrgh#10 @brentp
JOSS: statement of need.
One of the criteria for JOSS is:
```
A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
```

Please make this more explicit in the README and/or godocs

https://github.com/openjournals/joss-reviews/issues/517
