### 2018-01-08 kortschak/arrgh#10 @brentp
under the `Paper` section, the JOSS review template also lists:
```
 A statement of need: Do the authors clearly state what problems the software is
designed to solve and who the target audience is?
```
so adding a similar statement on audience should be added to the paper--though the paper already indicates the problems solved by the software.
