### 2018-01-08 kortschak/arrgh#9 @kortschak
I must have had those already installed from other projects. Maybe this is something for addition to the [OpenCPU Download and Install page](https://www.opencpu.org/download.html).

/cc @jeroen
