### 2018-01-08 df94e59
README: expand statement of need
