+++ paper/paper.md
@@ -23,6 +23,7 @@ The OpenCPU API provides a number of HTTP endpoints and a JSON Remote Procedure 
 Go is a simple statically typed compiled language that provides many benefits for scientific computing.
 However, Go currently lacks the rich statistical analysis tools available in R.
 So an interface between R and Go would allow building analytical tools utilising the simplicity and high performance of Go and the statistical analysis tools available within R.
+The arrgh package is intended to be used by Go developers wanting to add rich statistical analysis to Go analytical tools.
 
 The arrgh package provides a Go client for the OpenCPU system, including the management of a local server when needed.
 Data exchange between Go and OpenCPU is via standard HTTP content types, with arrgh providing tools to facilitate encoding and endpoint targeting.
