### 2018-01-11 #517 @brentp
@pjotrp I believe I have completed the review and the software meets the requirements as described with the inclusion of kortschack/arrgh#9
