### 2018-01-11 kortschak/arrgh#9 @kortschak
@brentp I'm trying to figure out exactly where the sys dependency comes in, and particularly the need to the `NO_APPARMOR` envvar; I have never needed to set this (also running on 16.04). Do you have the console output still from the install session you did?

I imagine the protobuf dependency comes in from the fact that opencpu can now do protobuf serialisation.
