### 2018-01-11 kortschak/arrgh#9 @kortschak
BTW Did you try to follow the instructions on the [OpenCPU download page](https://www.opencpu.org/download.html)?
