### 2018-01-11 kortschak/arrgh#9 @brentp
I just installed on another machine and the sys package installed without issue so I'm fine to close this iusse. I got this working in conda no problem by install r-base and then proceeding from there. There's also:
```
conda install -c quasiben r-opencpu
```

Even though I did install on my laptop, I tried to use `install.packages('opencpu')` so that I wouldn't be relying on something that wouldn't work on the cluster. The install.packages still requires protobuf which is not listed in that help.
