### 2018-01-11 kortschak/arrgh#9 @kortschak
OK. I'll wait for a while to see what happens with the issue at OpenCPU. If there is no movement there after a few days, I'll sort out what the deps are and add that list here. I'll ping you either way.
