### 2018-01-12 5389584
README: add link to CRAN page for OpenCPU
Also add link to debian source install instructions.
